Remove default layout (workaround for redirection pages)
All the converted wiki pages have an explicit layout (wiki, or
page if we want them in the sidebar), so this shouldn't have
much effect on our content. Except the index page.

This is a workaround for the jekyll-redirect-from plugin where
for the redirection pages the self-contained HTML snippet gets
embedded within the default layout (and thus does not work),
see #36.