Include tag/category links on wiki pages
e.g. Pages with the 'Cookbook' tag should have a link to
http://biopython.org/wiki/Category%3ACookbook

Note we have to replace spaces in the tags with underscores
so that e.g. 'Wiki Documentation' will link to the URL
http://peterjc.github.io/wiki/Category%3AWiki_Documentation

This is provisionally done as a bullet list above the page
title, and would benefit from some CSS styling, but is enough
to close #104.

Thank you to @Gasta88 for finding a related example which was
very helpful: http://charliepark.org/tags-in-jekyll/